@@ -4,7 +4,7 @@
   "description": "",
   "main": "index.js",
   "scripts": {
-    "build": "npm run lint && npm run clean && webpack --progress -p",
+    "build": "npm run lint && npm run clean && webpack --progress -p --config webpack.production.config.js",
     "start": "webpack-dev-server --progress --history-api-fallback --hot --inline --port 9000 --open",
     "dev": "npm run start",
     "dev:remote": "webpack-dev-server --progress --history-api-fallback --hot --inline --host 0.0.0.0 --port 9000 --open",
@@ -17,30 +17,35 @@
     "axios": "0.15.3",
     "classnames": "2.2.5",
     "dotenv": "4.0.0",
+    "express": "4.14.1",
+    "normalize.css": "^5.0.0",
     "react": "15.4.2",
     "react-dom": "15.4.2",
     "react-router": "3.0.2"
   },
   "devDependencies": {
-    "babel-core": "^6.7.0",
-    "babel-loader": "^6.2.4",
-    "babel-preset-es2015": "^6.6.0",
-    "babel-preset-react": "^6.5.0",
-    "copy-webpack-plugin": "^3.0.1",
-    "css-loader": "^0.25.0",
-    "eslint": "^3.4.0",
-    "eslint-config-airbnb": "^10.0.1",
-    "eslint-plugin-import": "^1.14.0",
-    "eslint-plugin-jsx-a11y": "^2.2.1",
-    "eslint-plugin-react": "^6.2.0",
-    "file-loader": "^0.9.0",
-    "html-webpack-plugin": "^2.9.0",
-    "postcss-cssnext": "^2.8.0",
-    "postcss-loader": "^0.11.1",
-    "rimraf": "^2.5.4",
-    "style-loader": "^0.13.0",
-    "url-loader": "^0.5.7",
-    "webpack": "^1.13.2",
-    "webpack-dev-server": "^1.15.1"
+    "babel-core": "6.22.1",
+    "babel-loader": "6.2.10",
+    "babel-preset-es2015": "6.22.0",
+    "babel-preset-react": "6.22.0",
+    "copy-webpack-plugin": "4.0.1",
+    "css-loader": "0.26.1",
+    "eslint": "3.15.0",
+    "eslint-config-airbnb": "14.1.0",
+    "eslint-plugin-import": "2.2.0",
+    "eslint-plugin-jsx-a11y": "4.0.0",
+    "eslint-plugin-react": "6.9.0",
+    "extract-text-webpack-plugin": "^2.0.0-rc.3",
+    "file-loader": "0.10.0",
+    "html-webpack-plugin": "2.28.0",
+    "postcss": "5.2.12",
+    "postcss-cssnext": "2.9.0",
+    "postcss-import": "9.1.0",
+    "postcss-loader": "1.3.0",
+    "rimraf": "2.5.4",
+    "style-loader": "0.13.1",
+    "url-loader": "0.5.7",
+    "webpack": "2.2.1",
+    "webpack-dev-server": "2.3.0"
   }
 }
