@@ -1,11 +1,11 @@
 import React from 'react';
-import { Router, Route, IndexRoute, browserHistory } from 'react-router';
+import { Router, Route, IndexRoute, hashHistory } from 'react-router';
 import Main from '../components/main';
 import Home from '../views/home';
 import About from '../views/about';
 
-const routes = (
-    <Router history={browserHistory} onUpdate={() => window.scrollTo(0, 0)}>
+const Routes = (
+    <Router history={hashHistory} onUpdate={() => window.scrollTo(0, 0)}>
         <Route path="/" component={Main}>
             <IndexRoute component={Home} />
             <Route path="about" component={About} />
@@ -13,4 +13,4 @@ const routes = (
     </Router>
 );
 
-module.exports = routes;
+export default Routes;
