@@ -1,8 +1,8 @@
 import React from 'react';
 
 const About = () => (
-    <div>
-        <h1>About</h1>
+    <div className="container">
+        <h2>About</h2>
     </div>
 );
 
